Example working with mock
