--- app/.babelrc
@@ -1,19 +1,0 @@
-{
-  "presets": [
-    [
-      "@nx/react/babel",
-      {
-        "runtime": "automatic"
-      }
-    ]
-  ],
-  "plugins": [
-    [
-      "styled-components",
-      {
-        "pure": true,
-        "ssr": true
-      }
-    ]
-  ]
-}
