+++ app/package.json
@@ -1,0 +1,29 @@
+{
+  "name": "example-app",
+  "private": true,
+  "version": "0.0.0",
+  "type": "module",
+  "scripts": {
+    "dev": "vite"
+  },
+  "dependencies": {
+    "@dxc-technology/halstack-react": "^16.0.0",
+    "@dxc-technology/halstack-react-hal": "^5.7.0",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "styled-components": "^5.0.1"
+  },
+  "devDependencies": {
+    "@types/react": "^18.3.12",
+    "@types/react-dom": "^18.3.1",
+    "@types/styled-components": "^5.1.34",
+    "@vitejs/plugin-react": "^4.3.3",
+    "typescript": "~5.6.2",
+    "util": "^0.12.5",
+    "vite": "^5.4.10",
+    "vite-plugin-node-polyfills": "^0.22.0"
+  },
+  "overrides": {
+    "esbuild": "0.25.1"
+  }
+}
