--- app/src/app/app.tsx
@@ -1,79 +1,0 @@
-import { HalTable, HalAutocomplete } from "@dxc-technology/halstack-react-hal";
-import {
-  DxcApplicationLayout,
-  DxcInset,
-  DxcHeading,
-  DxcFlex,
-} from "@dxc-technology/halstack-react";
-import React from "react";
-
-function App() {
-  return (
-    <DxcApplicationLayout>
-      <DxcApplicationLayout.Main>
-        <DxcInset space="2rem">
-          <DxcFlex direction="column" gap="4rem">
-            <DxcFlex direction="column" gap="2rem">
-              <DxcHeading level={2} text="HalTable example" />
-              <HalTable
-                mode="reduced"
-                collectionUrl={"http://localhost:3001/policies"}
-                columns={[
-                  {
-                    header: "Email",
-                    displayProperty: "email",
-                    sortProperty: "email",
-                  },
-                  {
-                    header: "Phone Number",
-                    displayProperty: "phone_number",
-                    sortProperty: "phone_number",
-                    onClickItemFunction: (value) => console.log(value),
-                  },
-                  {
-                    header: "Username",
-                    displayProperty: "username",
-                    sortProperty: "username",
-                    mapFunction: (item) => `test-${item.summary.username}`,
-                  },
-                ]}
-              />
-              <HalTable
-                collectionUrl={"http://localhost:3001/policies"}
-                columns={[
-                  {
-                    header: "Email",
-                    displayProperty: "email",
-                    sortProperty: "email",
-                  },
-                  {
-                    header: "Phone Number",
-                    displayProperty: "phone_number",
-                    sortProperty: "phone_number",
-                    onClickItemFunction: (value) => console.log(value),
-                  },
-                  {
-                    header: "Username",
-                    displayProperty: "username",
-                    sortProperty: "username",
-                    mapFunction: (item) => `test-${item.summary.username}`,
-                  },
-                ]}
-              />
-            </DxcFlex>
-            <DxcFlex direction="column" gap="2rem">
-              <DxcHeading level={2} text="HalAutocomplete example" />
-              <HalAutocomplete
-                collectionUrl="http://your-api/users"
-                propertyName="username"
-                label="Username"
-              />
-            </DxcFlex>
-          </DxcFlex>
-        </DxcInset>
-      </DxcApplicationLayout.Main>
-    </DxcApplicationLayout>
-  );
-}
-
-export default App;
