+++ package.json
@@ -14,9 +14,10 @@
     "build:watch": "tsup --watch",
     "dev": "cd app && npm run dev",
     "link": "npm link && cd app && npm link @dxc-technology/halstack-react-hal",
-    "test": "jest",
     "unlink": "cd app && npm unlink @dxc-technology/halstack-react-hal",
-    "copy": "node ./scripts/build/copy-readme.js"
+    "copy": "node ./scripts/build/copy-readme.js",
+    "test": "jest --collect-coverage",
+    "serve-server": "json-server src/tests/mockCollection.json & sleep 5"
   },
   "private": true,
   "dependencies": {
@@ -29,10 +30,10 @@
     "tslib": "^2.3.0"
   },
   "devDependencies": {
-    "@dxc-technology/halstack-react": "^15.0.0",
     "@babel/preset-env": "^7.20.0",
     "@babel/preset-react": "^7.20.0",
     "@babel/preset-typescript": "^7.20.0",
+    "@dxc-technology/halstack-react": "^15.0.0",
     "@testing-library/react": "14.0.0",
     "@types/jest": "^29.5.14",
     "@types/styled-components": "^5.1.34",
@@ -40,6 +41,7 @@
     "identity-obj-proxy": "^3.0.0",
     "jest": "^29.2.0",
     "jest-environment-jsdom": "^29.2.0",
+    "json-server": "^1.0.0-beta.3",
     "prettier": "3.3.3",
     "tsup": "^8.5.1",
     "typescript": "^5.6.3"
